Improve initial audio output route setting
This patch improves the initial audio output route setting using the
preferred device selected previously in the output selector from media
notification. Before, if headphones/headset was plugged in, we'd use it
even though the speaker is the preferred device.

If neither speaker or headphones/headset is the preferred device, than
we reset the preferred device to avoid desync of the route audio policy
since we'll default to speaker or headphones/headset (if present) when
not using audio session. This is because, the user could have selected
any device as the preferred output device (such as BT headset or USB
audio headset), but we can't seamlessly route to that device when using
FM device loopback.

While we're at it, replace the BluetoothAdapter with BluetoothManager to
retrieve the BT adapter, since that API has been deprecated.

Signed-off-by: iusmac <iusico.maxim@libero.it>