Unconditionally force use audio session when neither speaker nor headset is used
With this commit, we restore support for unconditionally switching from
FM device loopback to the audio session to play FM audio when neither a
speaker nor a wired headset is used (it could be a BT headset or USB
audio headset). This is because devices relying on the Audio HAL for FM
device loopback, most of time lack A2DP offload capability from the
vendor needed to process the FM output.

When attempting to route FM audio to SND_DEVICE_OUT_BT_A2DP (bt-a2dp)
device, the following is logged, indicating that the device is missing
the /vendor/lib64/liba2dpoffload.so library:

    D audio_hw_primary: SCO/A2DP is selected but they are not connected/ready hence dont route

Using audio session allows us to delegate FM audio processing to Android
that will send audio to any device.

Signed-off-by: iusmac <iusico.maxim@libero.it>